refactor: reduce controller and add exceptions

+++ src/main/java/com/kakao/cafe/web/service/ArticleService.java
@@ -2,6 +2,9 @@ package com.kakao.cafe.web.service;
 
 import com.kakao.cafe.domain.Article;
 import com.kakao.cafe.domain.Reply;
+import com.kakao.cafe.domain.Users;
+import com.kakao.cafe.web.exception.AccessDeniedException;
+import com.kakao.cafe.web.exception.UnauthorizedException;
 import com.kakao.cafe.web.repository.ArticleRepository;
 import com.kakao.cafe.web.repository.ReplyRepository;
 import org.springframework.stereotype.Component;
@@ -22,7 +25,8 @@ public class ArticleService {
         return articleRepository.selectAllArticles();
     }
 
-    public void addArticle(Article article) {
+    public void addArticle(Article article, Users author) {
+        article.setAuthor(author);
         articleRepository.insertArticle(article);
     }
 
@@ -30,7 +34,9 @@ public class ArticleService {
         return articleRepository.selectByArticleId(id);
     }
 
-    public void updateArticle(int id, Article updateArticle) {
+    public void updateArticle(int id, int userId, Article updateArticle) {
+        if (getByArticleId(id).getAuthor().getId() != userId)
+            throw new UnauthorizedException();
         if (updateArticle.getTitle().isBlank())
             throw new IllegalArgumentException("제목이 빈 값일 수 없습니다.");
         if (updateArticle.getContent().isBlank())
@@ -39,11 +45,20 @@ public class ArticleService {
     }
 
     public void deleteArticle(int id, int userId) {
+        if (getByArticleId(id).getAuthor().getId() != userId)
+            throw new AccessDeniedException();
         List<Reply> replies = articleRepository.selectByArticleId(id).getReplies();
         replies.forEach(reply -> {
             if (reply.getAuthor().getId() != userId) throw new IllegalArgumentException("글을 삭제할 수 없습니다.");
         });
         replies.forEach(reply -> replyRepository.deleteReply(reply.getId()));
         articleRepository.deleteArticle(id);
+    }
+
+    public Article getArticleUpdateForm(int id, int userId) {
+        Article article = getByArticleId(id);
+        if (article.getAuthor().getId() != userId)
+            throw new AccessDeniedException();
+        return article;
     }
 }
